requirements.txt
aiohttp>=3.8.0
aiofiles>=23.0.0
pyyaml>=6.0
docker>=6.0.0
requests>=2.28.0
click>=8.0.0
coloredlogs>=15.0
python-dateutil>=2.8.0
jsonschema>=4.0.0
psutil>=5.9.0
